+++ R/landscape.R
@@ -78,7 +78,10 @@ calculateLandscapeMetrics <- function(summaryPolys, polyCol, vtm, funList = NULL
       dplyr::summarise(
         N = length(value),
         mm = ifelse(N > 0, min(value, na.rm = TRUE), NA_real_),
+        q1 = ifelse(N > 0, quantile(value, 0.25, na.rm = TRUE), NA_real_),
+        md = ifelse(N > 0, median(value, na.rm = TRUE), NA_real_),
         mn = ifelse(N > 0, mean(value, na.rm = TRUE), NA_real_),
+        q3 = ifelse(N > 0, quantile(value, 0.75, na.rm = TRUE), NA_real_),
         mx = ifelse(N > 0, max(value, na.rm = TRUE), NA_real_),
         sd = ifelse(N > 0, sd(value, na.rm = TRUE), NA_real_),
         se = ifelse(N > 0, sd / sqrt(N), NA_real_),